Add users microposts RSS feed
